servicecheck: move code to own file
